refactor: 패키지 이름 변경 및 코드 컨벤션 정리
